Aula 3.6. Repository Supplier e Product

+++ services/product.service.js
@@ -1,11 +1,12 @@
 import productRepository from "../repositories/product.repository.js";
 import supplierRepository from "../repositories/supplier.repository.js";
+import saleRepository from '../repositories/sale.repository.js';
 
 async function createProduct(product) {
-    if (await supplierRepository.getSupplier(product.supplier_id)) {
+    if (await supplierRepository.getSupplier(product.supplierId)) {
         return await productRepository.insertProduct(product);
     } else {
-        throw new Error(`O supplier_id ${product.supplier_id} informado não existe`);
+        throw new Error(`O supplierId ${product.supplierId} informado não existe`);
     }
 }
 
@@ -18,14 +19,17 @@ async function getProduct(id) {
 }
 
 async function deleteProduct(id) {
+    const sale = saleRepository.getSalesByProductId(id);
+    if (sale) throw new Error("Não é possível excluir o produto pois ele já tem vendas.");
+
     await productRepository.deleteProduct(id);
 }
 
 async function updateProduct(product) {
-    if (await supplierRepository.getSupplier(product.supplier_id)) {
+    if (await supplierRepository.getSupplier(product.supplierId)) {
         return await productRepository.updateProduct(product);
     } else {
-        throw new Error(`O supplier_id ${product.supplier_id} informado não existe`);
+        throw new Error(`O supplierId ${product.supplierId} informado não existe`);
     }
 }
